requirements.txt
# MJSRecon - Modular JavaScript Reconnaissance Tool
# Python Dependencies

# Core HTTP and networking
requests>=2.31.0
urllib3>=2.0.0
aiohttp>=3.8.0

# Progress bars and UI
tqdm>=4.65.0
rich>=13.0.0

# JavaScript parsing and beautification
jsbeautifier>=1.14.0

# HTML/XML parsing
lxml>=4.9.0

# File handling and requests with file:// protocol
requests-file>=1.5.2

# Configuration management
PyYAML>=6.0

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Security and cryptography (for secret detection)
cryptography>=41.0.0

# Additional utilities for enhanced functionality
colorama>=0.4.6
click>=8.1.0

# Performance monitoring
psutil>=5.9.0 
uro 